fix loading components
