+++ src/components/cards/ServiceCard.jsx
@@ -17,7 +17,7 @@ const ServiceCard = ({service}) =>{
                     <div className="flex items-center flex-wrap ">
                     {user ? (
                         <>
-                            <label htmlFor="cita" className="cursor-pointer text-green-400 inline-flex font-semibold items-center md:mb-2 lg:mb-0">Agendar cita
+                            <label htmlFor={service.id} className="cursor-pointer text-green-400 inline-flex font-semibold items-center md:mb-2 lg:mb-0">Agendar cita
                                 <svg className="w-4 h-4 ml-2" viewBox="0 0 24 24" stroke="currentColor" strokeWidth="2" fill="none" strokeLinecap="round" strokeLinejoin="round">
                                     <path d="M5 12h14"></path>
                                     <path d="M12 5l7 7-7 7"></path>
