+++ src/components/tables/OrdersTables.jsx
@@ -1,12 +1,35 @@
-import { useGetOrdersByClientQuery } from "../../store/serverApi";
 import OrderElement from "./OrderElement";
 import Loader from "../main/Loader"
 import { useSelector } from "react-redux";
+import axios from "axios";
+import { useEffect, useState } from "react";
 
 const OrdersTables =({setAction})=>{
 
     const id = useSelector((state) => state.users.currentUser);
-    const {data: ordenes,isSuccess,isLoading} = useGetOrdersByClientQuery(id)
+    const [ordenes,setOrdenes] = useState([])
+    const [isLoading,setIsLoading] = useState(true)
+
+    useEffect(()=>{
+        getOrders(id)
+    },[id])
+
+    useEffect(() => {
+        setIsLoading(true)
+        setTimeout(() => {
+            setIsLoading(false)
+        }, 2000);
+    },[])
+
+    const getOrders = async (id)=>{
+        await axios.get(`https://order-production-bfbc.up.railway.app/api/cliente/order/client/${id}`)
+          .then(response => {
+            setOrdenes(response.data)
+          })
+          .catch(error => {
+            console.log(error)
+          })
+    }
 
     const modal = (id,nombre) =>{
        setAction(id,nombre)
@@ -20,9 +43,8 @@ const OrdersTables =({setAction})=>{
                         <div className=" min-w-0 mb-6 break-words border-0 border-transparent border-solid shadow-soft-xl rounded-2xl bg-clip-border ">
                             <div className="px-6 py-4 mb-0 bg-white border-b-0 border-b-solid rounded-2xl h-3gl">
                                 <div className="px-4 overflow-x-auto md:overflow-y-auto md:max-h-[30rem]">
-                                {isLoading && <Loader />}
-                                {isSuccess && (
-                                   ordenes.length !== 0 ? (
+                                {isLoading ? <Loader /> : (
+                                    ordenes !== null ? (
                                         <table className=" items-center w-full mb-0 align-top border-gray-200 text-slate-500">
                                             <thead className="align-bottom">
                                                 <tr>
@@ -44,7 +66,7 @@ const OrdersTables =({setAction})=>{
                                     ):(
                                         <p className="mt-3 text-center">No existen órdenes asociadas</p>
                                     )
-                                )}    
+                                )}   
                                 </div>
                             </div>
                         </div>
